datasheetminer/requirements.txt
boto3>=1.39.10
google-genai>=1.26.0
python-dotenv>=1.1.1
requests>=2.32.4
httpx>=0.27.0
awslambdaric>=0.2.1
